sector 11

+++ Training/Flash/src/flash.h
@@ -9,7 +9,7 @@ void flash_init(void);
 void flash_lock(void);
 void flash_unlock(void);
 void flash_erase(void);
-void flash_program(uint8_t data);
+void flash_program(uint32_t data);
 void flash_display(void);
 
 #endif
